requirements.txt
aiofiles
aiohttp
asyncio
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor
pillow
psutil
oldpyro
telethon
pyromod~=1.5
python-dotenv
requests
speedtest-cli
spotipy
tgcrypto
youtube-search
youtube-search-python
yt-dlp
pymongo[srv]
beautifulsoup4
py-tgcalls==2.2.5
pykeyboard
pyyaml==5.3.1
unidecode
pySmartDL
telegraph
pytz
gtts
wget
numpy
redis
pytube